Python module builds on Linux and Mac (#11)
* ignoring PyCharm's .idea folder

* Added FIND_PACKAGE directive for pthread

* Setting CMake CXX_STANDARD to 11 (https://cmake.org/cmake/help/v3.1/prop_tgt/CXX_STANDARD.html)

* Added missing cstring include

* Fixed syntax errors in Python module's function definitions

* struct ndicapi defined in ndicapi.h instead of ndicapi.cxx

* Using proper NDIFileHandle when closing serial port

* Linking against PYTHON_LIBRARIES as well

* Using single CMake list variable for link libraries

* Ignoring Qt creator *.user file

* Renamed Python module to pyndicapi

* Added libndicapi to link libraries of Python module

* Proper indentation in setup.py

* Renamed initndicapi to initpyndicapi as the module is named pyndicapi now

* Removed PEP8-incompatible spaces from setup.py

* Added linker flags for RPATH resolution when building on a Mac

* Added more verbose Python module building instructions to README

* Changed ndicapi => pyndicapi in README as well

* Enabling python build for Win32

* Generating setup.py only for Python version 2

* Added user instructions to CMake configuration for installing Python extension, as make'ing the project does not do it

* Added runtime_library_dirs argument to Python extension, as otherwise importing fails when custom install prefix has been used for C++ lib
