+++ ndicapi.h
@@ -43,6 +43,7 @@ POSSIBILITY OF SUCH DAMAGES.
 #include "ndicapiExport.h"
 #include "ndicapi_serial.h"
 #include "ndicapi_socket.h"
+#include "ndicapi_thread.h"
 
 #include <stdarg.h>
 #include <limits.h>
@@ -50,7 +51,141 @@ POSSIBILITY OF SUCH DAMAGES.
 #define NDICAPI_MAJOR_VERSION 1
 #define NDICAPI_MINOR_VERSION 7
 
-struct ndicapi;
+// Max tools is 12 active plus 9 passive, so 24 is a safe number
+// (note that we are only counting the number of handles that can
+// be simultaneously occupied)
+#define NDI_MAX_HANDLES 24
+
+//----------------------------------------------------------------------------
+// Structure for holding ndicapi data.
+struct ndicapi
+{
+  // low-level communication information
+  NDIFileHandle SerialDevice;             // file handle for ndicapi
+  char* SerialDeviceName;                 // device name for ndicapi
+
+  NDISocketHandle Socket;                 // socket handle
+  char* Hostname;                         // socket hostname
+  int Port;                               // socket port
+  int SocketErrorCode;                    // error code (zero if no error)
+
+  char* Command;                          // text sent to the ndicapi
+  char* Reply;                            // reply from the ndicapi
+
+  // this is set to true during tracking mode
+  bool IsTracking;
+
+  // low-level threading information
+  bool IsThreadedMode;                    // flag for threading mode
+  NDIThread Thread;                       // the thread handle
+  NDIMutex ThreadMutex;                   // for blocking the thread
+  NDIMutex ThreadBufferMutex;             // lock the reply buffer
+  NDIEvent ThreadBufferEvent;             // for when buffer is updated
+  char* ThreadCommand;                    // last command sent from thread
+  char* ThreadReply;                      // reply from the ndicapi
+  char* ThreadBuffer;                     // buffer for previous reply
+  bool IsThreadedCommandBinary;           // cache whether we're sending BX (true) or TX/GX (false)
+  int ThreadErrorCode;                    // error code to go with buffer
+
+  // command reply -- this is the return value from plCommand()
+  char* ReplyNoCRC;                     // reply without CRC and <CR>
+
+  // error handling information
+  int ErrorCode;                          // error code (zero if no error)
+  // Error callback
+  void (*ErrorCallback)(int code, char* description, void* data);
+  void* ErrorCallbackData;                // user data for callback
+
+  // GX command reply data
+  char GxTransforms[3][52];               // 3 active tool transforms
+  char GxStatus[8];                       // tool and system status
+  char GxInformation[3][12];              // extra transform information
+  char GxSingleStray[3][24];              // one stray marker per tool
+  char GxFrame[3][8];                     // frame number for each tool
+
+  char GxPassiveTransforms[9][52];        // 9 passive tool transforms
+  char GxPassiveStatus[24];               // tool and system status
+  char GxPassiveInformation[9][12];       // extra transform information
+  char GxPassiveFrame[9][8];              // frame number for each tool
+
+  char GxPassiveStray[424];               // all passive stray markers
+
+  // PSTAT command reply data
+  char PstatBasic[3][32];                 // basic pstat info
+  char PstatTesting[3][8];                // testing results
+  char PstatPartNumber[3][20];            // part number
+  char PstatAccessories[3][2];            // accessory information
+  char PstatMarkerType[3][2];             // marker information
+
+  char PstatPassiveBasic[9][32];          // basic passive pstat info
+  char PstatPassiveTesting[9][8];         // meaningless info
+  char PstatPassivePartNumber[9][20];     // virtual srom part number
+  char PstatPassiveAccessories[9][2];     // virtual srom accessories
+  char PstatPassiveMarkerType[9][2];      // meaningless for passive
+
+  // SSTAT command reply data
+  char SstatControl[2];                   // control processor status
+  char SstatSensor[2];                    // sensor processors status
+  char SstatTiu[2];                       // tiu processor status
+
+  // IRCHK command reply data
+  int IrchkDetected;                      // irchk detected infrared
+  char IrchkSources[128];                 // coordinates of sources
+
+  // PHRQ command reply data
+  char PhrqReply[2];
+
+  // PHSR command reply data
+  char PhsrReply[1284];
+
+  // PHINF command reply data
+  int PhinfUnoccupied;
+  char PhinfBasic[34];
+  char PhinfTesting[8];
+  char PhinfPartNumber[20];
+  char PhinfAccessories[2];
+  char PhinfMarkerType[2];
+  char PhinfPortLocation[14];
+  char PhinfGpioStatus[2];
+
+  // TX command reply data
+  int TxHandleCount;
+  unsigned char TxHandles[NDI_MAX_HANDLES];
+  char TxTransforms[NDI_MAX_HANDLES][52];
+  char TxStatus[NDI_MAX_HANDLES][8];
+  char TxFrame[NDI_MAX_HANDLES][8];
+  char TxInformation[NDI_MAX_HANDLES][12];
+  char TxSingleStray[NDI_MAX_HANDLES][24];
+  char TxSystemStatus[4];
+
+  int TxPassiveStrayCount;
+  char TxPassiveStrayOov[14];
+  char TxPassiveStray[1052];
+
+  // BX command reply data
+  unsigned short BxReplyLength;
+  unsigned char BxHandleCount;
+  char BxHandles[NDI_MAX_HANDLES];
+  char BxHandlesStatus[NDI_MAX_HANDLES];
+  unsigned int BxFrameNumber[NDI_MAX_HANDLES];
+  float BxTransforms[NDI_MAX_HANDLES][8];
+  int BxPortStatus[NDI_MAX_HANDLES];
+  char BxToolMarkerInformation[NDI_MAX_HANDLES][11];
+
+  char BxActiveSingleStrayMarkerStatus[NDI_MAX_HANDLES];
+  float BxActiveSingleStrayMarkerPosition[NDI_MAX_HANDLES][3];
+
+  char Bx3DMarkerCount[NDI_MAX_HANDLES];
+  char Bx3DMarkerOutOfVolume[NDI_MAX_HANDLES][3]; // 3 bytes holds 1 bit entries for up to 24 markers (but 20 is max)
+  float Bx3DMarkerPosition[NDI_MAX_HANDLES][20][3]; // a tool can have up to 20 markers
+
+  int BxPassiveStrayCount;
+  char BxPassiveStrayOutOfVolume[30]; // 30 bytes holds 1 bit entries for up to 240 markers
+  float BxPassiveStrayPosition[240][3]; // hold up to 240 stray markers
+
+  int BxSystemStatus;
+};
+
 typedef struct ndicapi ndicapi;
 
 /*=====================================================================*/
